flask_app/requirements.txt
docker==7.1.0
Flask==3.0.3
python-dotenv==1.0.0
